+++ tests/schema.test.ts
@@ -1,12 +1,12 @@
 import { schema } from "src/schema";
 
 describe("schema", () => {
-  it("returns an async validator function", () => {
+  it("returns an async validator function", async () => {
     const validator = schema({}, {});
-    expect(validator({})).toEqual({});
+    expect(await validator({})).toEqual({});
   });
 
-  it("runs validations", () => {
+  it("runs validations", async () => {
     const validator = schema(
       {
         foo: () => ({ ok: true }),
@@ -25,14 +25,33 @@ describe("schema", () => {
       }
     );
 
-    expect(validator({})).toEqual({
+    expect(await validator({})).toEqual({
       bar: "Value: ''; Args: 1 2",
       bux: "Failure!"
     });
 
-    expect(validator({ foo: "abc", bar: "bux", unrelated: "?" })).toEqual({
-      bar: "Value: 'bux'; Args: 1 2",
-      bux: "Failure!"
+    expect(await validator({ foo: "abc", bar: "bux", unrelated: "?" })).toEqual(
+      {
+        bar: "Value: 'bux'; Args: 1 2",
+        bux: "Failure!"
+      }
+    );
+  });
+
+  it("accepts async validators", async () => {
+    const validator = schema(
+      {
+        foo: async () => ({ ok: false, name: "bread", args: ["a", "b"] })
+      },
+      {
+        foo: {
+          bread: (value, arg1, arg2) => `Value: ${value}; Args: ${arg1} ${arg2}`
+        }
+      }
+    );
+
+    expect(await validator({ foo: "bar" })).toEqual({
+      foo: "Value: bar; Args: a b"
     });
   });
 });